# Conflicts:
#	server/src/main/java/org/tctalent/server/repository/db/CandidateSpecification.java
#	server/src/main/java/org/tctalent/server/service/db/BackgroundProcessingService.java
#	server/src/main/java/org/tctalent/server/service/db/es/ElasticsearchServiceImpl.java
#	server/src/main/java/org/tctalent/server/service/db/impl/BackgroundProcessingServiceImpl.java
#	server/src/main/java/org/tctalent/server/service/db/impl/CandidateServiceImpl.java
#	server/src/main/java/org/tctalent/server/util/PersistenceContextHelper.java
#	server/src/main/resources/db/migration/V1_323__drop_zombie_cvjc_fields.sql
#	server/src/main/resources/db/migration/V1_324__create_table_duolingo_coupons.sql
#	server/src/main/resources/db/migration/V1_325__add_potential_duplicate_to_candidate_and_saved_search.sql
#	ui/admin-portal/src/app/app.module.ts
#	ui/admin-portal/src/app/components/candidates/potential-duplicates/duplicates-detail/duplicates-detail.component.html
#	ui/admin-portal/src/app/components/candidates/potential-duplicates/duplicates-detail/duplicates-detail.component.spec.ts
#	ui/admin-portal/src/app/components/candidates/potential-duplicates/duplicates-detail/duplicates-detail.component.ts
#	ui/admin-portal/src/app/components/candidates/potential-duplicates/potential-duplicate-icon/potential-duplicate-icon.component.html
#	ui/admin-portal/src/app/components/candidates/potential-duplicates/potential-duplicate-icon/potential-duplicate-icon.component.spec.ts
#	ui/admin-portal/src/app/components/candidates/potential-duplicates/potential-duplicate-icon/potential-duplicate-icon.component.ts
#	ui/admin-portal/src/app/components/candidates/show/show-candidates.component.ts
#	ui/admin-portal/src/app/components/candidates/view/view-candidate.component.ts
#	ui/admin-portal/src/app/components/job/view/info/view-job-info/view-job-info.component.html
#	ui/admin-portal/src/app/components/job/view/view-job/view-job.component.html
#	ui/admin-portal/src/app/components/search/define-search/define-search.component.html
#	ui/admin-portal/src/app/components/search/define-search/define-search.component.ts
#	ui/admin-portal/src/app/components/util/candidate-search-card/candidate-search-card.component.ts
#	ui/admin-portal/src/app/services/authorization.service.ts
#	ui/admin-portal/src/app/services/candidate.service.ts
#	ui/admin-portal/src/styles.scss

+++ server/src/main/java/org/tctalent/server/service/db/impl/CandidateStatsServiceImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Talent Beyond Boundaries.
+ * Copyright (c) 2024 Talent Catalog.
  *
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU Affero General Public License as published by the Free
@@ -16,15 +16,15 @@
 
 package org.tctalent.server.service.db.impl;
 
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import jakarta.persistence.Query;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.lang.Nullable;
@@ -735,7 +735,7 @@ public class CandidateStatsServiceImpl implements CandidateStatsService {
         for (Object obj: objects) {
             Object[] row = (Object[]) obj;
             String label = row[0] == null ? "undefined" : row[0].toString();
-            DataRow dataRow = new DataRow(label, (BigInteger)row[1]);
+            DataRow dataRow = new DataRow(label, BigInteger.valueOf((Long)row[1]));
             dataRows.add(dataRow);
         }
         return dataRows;
